Rename package to io.victoriametrics
