+++ pom.xml
@@ -8,7 +8,7 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>com.victoriametrics.client</groupId>
+    <groupId>io.victoriametrics.client</groupId>
     <artifactId>client-jvm</artifactId>
     <version>1.0-SNAPSHOT</version>
