--- src/main/java/com/victoriametrics/client/validator/InvalidMetricNameException.java
@@ -1,8 +1,0 @@
-package com.victoriametrics.client.validator;
-
-public class InvalidMetricNameException extends RuntimeException {
-
-    public InvalidMetricNameException(String message) {
-        super(message);
-    }
-}
