+++ src/main/java/io/victoriametrics/client/metrics/Gauge.java
@@ -1,4 +1,4 @@
-package com.victoriametrics.client.metrics;
+package io.victoriametrics.client.metrics;
 
 import java.util.Objects;
 import java.util.function.Supplier;
